+++ packages/rnv/src/engineTools/engine-rn-electron-next/index.js
@@ -181,6 +181,9 @@ export const _taskLog = async (c) => {
 };
 TASKS[TASK_LOG] = _taskLog;
 
+const applyTemplate = async () => true;
+
 export default {
-    runTask
+    runTask,
+    applyTemplate
 };
