+++ packages/rnv/src/engineTools/engine-rn-next/index.js
@@ -181,6 +181,9 @@ const runTask = async (c, task) => {
     return TASKS[task](c);
 };
 
+const applyTemplate = async () => true;
+
 export default {
-    runTask
+    runTask,
+    applyTemplate
 };
