+++ packages/rnv/src/engineTools/engine-rn-web/index.js
@@ -293,6 +293,9 @@ const runTask = async (c, task) => {
     return TASKS[task](c);
 };
 
+const applyTemplate = async () => true;
+
 export default {
-    runTask
+    runTask,
+    applyTemplate
 };
