+++ packages/rnv/src/engineTools/index.js
@@ -1,7 +1,9 @@
 /* eslint-disable import/no-cycle */
-import { logError, logRaw } from '../systemTools/logger';
+import { logDebug, logTask } from '../systemTools/logger';
+import { isPlatformSupported } from '../platformTools';
 import {
-    getConfigProp
+    getConfigProp,
+    isBuildSchemeSupported
 } from '../common';
 
 import EngineRn from './engine-rn';
@@ -18,6 +20,17 @@ const ENGINES = {
     'engine-rn-electron-next': EngineRnElectronNext
 };
 
+const EngineNoOp = {
+    runTask: async (c, task) => {
+        logTask('EngineNoOp:runTask');
+        await isPlatformSupported(c);
+        await isBuildSchemeSupported(c);
+        return getEngineRunner(c).runTask(c, task);
+    },
+    applyTemplate: () => {
+    }
+};
+
 
 export const getEngineByPlatform = (c, platform, ignoreMissingError) => {
     let selectedEngineKey;
@@ -25,8 +38,8 @@ export const getEngineByPlatform = (c, platform, ignoreMissingError) => {
         selectedEngineKey = getConfigProp(c, platform, 'engine');
         const selectedEngine = c.files.rnv.engines.config?.engines?.[selectedEngineKey];
         if (!selectedEngine && !ignoreMissingError) {
-            logError(`Engine: ${selectedEngineKey} does not exists or is not registered`);
-            logRaw(new Error());
+            logDebug(`ERROR: Engine: ${selectedEngineKey} does not exists or is not registered ${new Error()}`);
+            // logRaw(new Error());
         }
         return selectedEngine;
     }
@@ -36,10 +49,15 @@ export const getEngineByPlatform = (c, platform, ignoreMissingError) => {
 
 export const getEngineRunner = (c) => {
     const selectedEngine = getEngineByPlatform(c, c.platform);
+    if (!selectedEngine) {
+        return EngineNoOp;
+        // throw new Error(`Cound not find engine with ${chalk.white(c.platform)} platform support`);
+    }
     const engine = ENGINES[selectedEngine?.id];
     if (!engine) {
-        logError(`Cound not find active engine with id ${selectedEngine?.id}. Available engines:
-${Object.keys(ENGINES).join(', ')}`);
+        return EngineNoOp;
+        //         throw new Error(`Cound not find active engine with id ${selectedEngine?.id}. Available engines:
+        // ${Object.keys(ENGINES).join(', ')}`);
     }
     return engine;
 };
